fix(tui): scroll the detail pane past line 65,535
The detail scroll offset was a u16 and the panels scrolled through Paragraph's u16 offset, so no row past line 65,535 could be shown. The offset is now a usize end to end: the coefficient views build only the rows in view (no placeholder lines for the rest), the rows left out stand in for the scroll, and what remains is dropped from the front rather than scrolled by a u16. The summary and raw-text views take the wider offset the same way.

Dropping the placeholders also takes the per-frame cost of the modified views off the model size. render bench, criterion vs the pre-windowing baseline:
  diff_objective_modified_200k   1.94 ms -> 0.51 ms  (-73.9%, p < 0.05)
  diff_constraint_modified_200k  1.58 ms -> 0.10 ms  (-93.7%, p < 0.05)

+++ tui/src/state.rs
@@ -407,7 +407,7 @@ pub struct JumpEntry {
     /// index: re-sorting, `ignore_order`, a tolerance change or a watch reload
     /// all move entries, and the jump must land on the same one.
     pub entry_name: Option<String>,
-    pub detail_scroll: u16,
+    pub detail_scroll: usize,
     pub filter: DiffFilter,
 }
